+++ andhow-system-tests/src/test/java/org/yarnandtail/andhow/AndHow_AliasInTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import org.springframework.mock.jndi.SimpleNamingContextBuilder;
 import org.yarnandtail.andhow.api.AppFatalException;
 import org.yarnandtail.andhow.internal.ConstructionProblem;
-import org.yarnandtail.andhow.load.JndiLoader;
+import org.yarnandtail.andhow.load.std.JndiLoader;
 import org.yarnandtail.andhow.name.CaseInsensitiveNaming;
 import org.yarnandtail.andhow.property.IntProp;
 import org.yarnandtail.andhow.property.StrProp;
