+++ andhow-system-tests/src/test/java/org/yarnandtail/andhow/example/restclient/SampleRestClientAppTest.java
@@ -13,7 +13,7 @@ import org.yarnandtail.andhow.api.AppFatalException;
 import org.yarnandtail.andhow.api.Property;
 import org.yarnandtail.andhow.internal.LoaderProblem;
 import org.yarnandtail.andhow.internal.ValueProblem;
-import org.yarnandtail.andhow.load.JndiLoader;
+import org.yarnandtail.andhow.load.std.JndiLoader;
 import org.yarnandtail.andhow.load.PropertyFileOnClasspathLoader;
 import org.yarnandtail.andhow.load.KeyValuePairLoader;
