+++ andhow/src/main/java/org/yarnandtail/andhow/load/std/JndiLoader.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.load;
+package org.yarnandtail.andhow.load.std;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,6 +6,7 @@ import javax.naming.*;
 import org.yarnandtail.andhow.GroupInfo;
 import org.yarnandtail.andhow.api.*;
 import org.yarnandtail.andhow.internal.LoaderProblem.JndiContextLoaderProblem;
+import org.yarnandtail.andhow.load.BaseLoader;
 import org.yarnandtail.andhow.property.QuotedSpacePreservingTrimmer;
 import org.yarnandtail.andhow.property.StrProp;
 import org.yarnandtail.andhow.sample.JndiLoaderSamplePrinter;
