+++ andhow/src/main/java/org/yarnandtail/andhow/load/std/StdFixedValueLoader.java
@@ -1,0 +1,11 @@
+package org.yarnandtail.andhow.load.std;
+
+import org.yarnandtail.andhow.load.FixedValueLoader;
+
+/**
+ *
+ * @author ericeverman
+ */
+public class StdFixedValueLoader extends FixedValueLoader {
+
+}
