+++ andhow/src/test/java/org/yarnandtail/andhow/load/std/EnviromentVariableLoaderTest.java
@@ -1,5 +1,6 @@
-package org.yarnandtail.andhow.load;
+package org.yarnandtail.andhow.load.std;
 
+import org.yarnandtail.andhow.load.std.EnviromentVariableLoader;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
