+++ andhow/src/test/java/org/yarnandtail/andhow/load/std/SystemPropertyLoaderTest.java
@@ -1,5 +1,6 @@
-package org.yarnandtail.andhow.load;
+package org.yarnandtail.andhow.load.std;
 
+import org.yarnandtail.andhow.load.std.SystemPropertyLoader;
 import org.yarnandtail.andhow.util.AndHowUtil;
 import org.junit.After;
